Refactor star system generation logic and update corresponding files
Revamped the star system generation logic in the project, which involved removing obsolete files and adding new ones with updated implementations. The changes include the deletion of OrbitNumber.scala and old versions of SystemGenerator.scala, and the creation of OrbitsGenerator.scala, StarGenerator.scala, Orbit.scala and GenUtils.scala. All the logic related to OrbitNumbers is now dealt with in the new classes, improving the modularity and clarity of the code. Additionally, a potential scalafix tool addition was commented out in the plugins.sbt file.